Remove `homebrew` submodule,
