+++ project/BuildPlugin.scala
@@ -311,133 +311,3 @@ object Header {
       |   ***********************************************************
     """.stripMargin
 }
-
-private object GitUtils {
-
-  import org.eclipse.jgit.api.Git
-  import org.eclipse.jgit.storage.file.FileRepositoryBuilder
-  import org.eclipse.jgit.lib.ObjectId
-
-  def latestCommitIn(dir: sbt.File): String = {
-    val repository =
-      new FileRepositoryBuilder().setGitDir(dir).readEnvironment().findGitDir().build()
-    val head = repository.resolve("HEAD")
-    ObjectId.toString(head)
-  }
-
-  def commitChangesIn(dir: sbt.File, changes: Seq[String], message: String): Unit = {
-    val git = Git.open(dir)
-    val add = git.add()
-    val cmd = changes.foldLeft(git.add) {
-      case (cmd, path) => cmd.addFilepattern(path)
-    }
-    cmd.call()
-    git.commit.setMessage(message).call()
-  }
-
-  def latestTagIn(dir: sbt.File): String = {
-    val git = Git.open(dir)
-    Option(git.describe().call()).getOrElse("no-tag-yet")
-  }
-
-  def submoduleDir(name: String) = Def.setting {
-    BuildKeys.buildBase.value / ".git" / "modules" / name
-  }
-
-}
-
-private object ReleaseUtils {
-  val installScript = Def.setting { BuildKeys.buildBase.value / "bin" / "install.py" }
-  val nailgunCommit = Def.setting {
-    GitUtils.latestCommitIn(GitUtils.submoduleDir("nailgun").value)
-  }
-  val versionedInstallScript = Def.task {
-    IO.readLines(installScript.value) match {
-      case shebang :: rest =>
-        val scriptTarget = Keys.target.value / "install.py"
-        val customizedVariables =
-          List(s"""NAILGUN_COMMIT = "${nailgunCommit.value}"""",
-               s"""BLOOP_VERSION = "${Keys.version.value}"""")
-        val newContent = shebang :: customizedVariables ++ rest
-        IO.writeLines(scriptTarget, newContent)
-        scriptTarget
-      case _ =>
-        sys.error(installScript.value.getAbsolutePath + " was empty?")
-    }
-  }
-  val writeHomebrewFormula = Def.task {
-    val base = BuildKeys.buildBase.value
-    val formulaTarget = base / "homebrew" / "bloop.rb"
-    val tagName = GitUtils.latestTagIn(base)
-    val installSha = sha256(versionedInstallScript.value)
-    val content =
-      s"""class Bloop < Formula
-         |  desc "Bloop gives you fast edit/compile/test workflows for Scala."
-         |  homepage "https://github.com/scalacenter/bloop"
-         |  version "${Keys.version.value}"
-         |  url "https://github.com/scalacenter/bloop/releases/download/${tagName}/install.py"
-         |  sha256 "${installSha}"
-         |  bottle :unneeded
-         |
-         |  depends_on "python"
-         |  depends_on :java => "1.8+"
-         |
-         |  def install
-         |      mkdir "bin"
-         |      system "python2", "install.py", "--dest", "bin", "--version", version
-         |      prefix.install "bin"
-         |  end
-         |
-         |  def plist; <<~EOS
-         |<?xml version="1.0" encoding="UTF-8"?>
-         |<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-         |<plist version="1.0">
-         |<dict>
-         |    <key>Label</key>
-         |    <string>#{plist_name}</string>
-         |    <key>ProgramArguments</key>
-         |    <array>
-         |        <string>#{bin}/bloop-server</string>
-         |    </array>
-         |    <key>KeepAlive</key>
-         |    <true/>
-         |</dict>
-         |</plist>
-         |          EOS
-         |      end
-         |
-         |  test do
-         |  end
-         |end""".stripMargin
-    IO.write(formulaTarget, content)
-  }
-  val updateHomebrewFormula = Def.task {
-    writeHomebrewFormula.value
-    val buildBase = BuildKeys.buildBase.value
-    val homebrewGitDir = GitUtils.submoduleDir("homebrew").value
-    val homebrewBase = buildBase / "homebrew"
-    val changed = "bloop.rb" :: Nil
-    GitUtils.commitChangesIn(homebrewGitDir,
-                             changed,
-                             "Update formula to Bloop " + Keys.version.value)
-  }
-
-  private def sha256(file: sbt.File): Option[String] = {
-    import java.nio.file.Files
-    import java.security.MessageDigest
-    scala.util.Try {
-      val digest = MessageDigest.getInstance("SHA-256")
-      val bytes = Files.readAllBytes(file.toPath)
-      val hash = digest.digest(bytes)
-      val hexString = new StringBuilder()
-      hash.foreach { byte =>
-        val hex = Integer.toHexString(0xff & byte)
-        if (hex.length == 1) hexString.append('0')
-        else hexString.append(hex)
-      }
-
-      hexString.toString
-    }.toOption
-
-  }
-}
